Eliminate dependency on libsodium - use OpenSSL's RNG.

+++ include/ntp_random.h
@@ -1,1 +1,1 @@
-long ntp_random (void);
+int32_t ntp_random (void);
